+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -243,3 +243,4 @@
 #import "SentryCrashCachedData.h"
 #import "SentryCrashInstallation+Private.h"
 #import "SentryCrashMonitor_MachException.h"
+#import "SentrySessionReplaySyncC.h"